fix: refactor Landing Components for readability
