+++ README.md
@@ -1,4 +1,5 @@
 <a name="readme-top"></a>
+
 <p align="center">
     <a href="https://github.com/Dcom-KHU/dcom-landing-page/issues">
         <img alt="Issues" src="https://img.shields.io/github/issues/Dcom-KHU/dcom-landing-page">
@@ -9,7 +10,6 @@
         <a href="https://github.com/Dcom-KHU/dcom-landing-page/#contributors">
             <img alt="Issues" src="https://img.shields.io/badge/all_contributors-1-orange.svg?style=flat-square">
                 </a>
-
 
 </p>
 <!-- PROJECT LOGO -->
@@ -32,40 +32,49 @@
  <img src="https://img.shields.io/badge/JavaScript-F7DF1E?style=flat-square&logo=javascript&logoColor=black"/>
   <img src="https://img.shields.io/badge/github%20actions-%232671E5.svg?style=flat-square&logo=githubactions&logoColor=white"/>
 
-
   </p>
 
 </div>
 
 ## Contribution Guide
+
 해당 프로젝트는 여러분들의 **contribution을 적극 환영**합니다!  
 블로그 글이나 프로젝트 추가도 좋고, 기능이나 디자인 개선도 좋습니다. 자세한 내용은 **아래 Guide를 참고**해주세요!
-### Basic Guide  
 
-1. Fork the repository  
-2. Clone the repository  
-3. Create a new branch to hold your change  
-``` bash
+### Basic Guide
+
+1. Fork the repository
+2. Clone the repository
+3. Create a new branch to hold your change
+
+```bash
 git checkout -b your-branch-name
 ```
-⚠️**Do not work** on the `main` branch⚠️  
+
+⚠️**Do not work** on the `main` branch⚠️
 
 4. Setting guide
-``` bash
+
+```bash
 npm install
 npm start
 ```
+
 5.  Push your change to remote branch
+
 ```
 git push -u origin your-branch-name
 ```
+
 6.  Make a full request to reflect your changes.
+
 ### Add item
+
 - `/posts` 에 **블로그 글**을 추가하기 위해서는
-`content/posts/posts.json` 에서 내용을 추가해주세요!
+  `content/posts/posts.json` 에서 내용을 추가해주세요!
 
 - `/projects` 에 **프로젝트**를 추가하기 위해서는
-`content/posts/projects.json` 에서 내용을 추가해주세요!
+  `content/posts/projects.json` 에서 내용을 추가해주세요!
 
 ## Contributors
 
@@ -92,6 +101,8 @@ git push -u origin your-branch-name
 
 <!-- ALL-CONTRIBUTORS-LIST:END -->
 
-## Contact  
+## Contact
+
 프로젝트에 궁금한 점이 있으면 언제나 아래 연락처로 연락주세요.
+
 - Yongwoo Song, ywsong.dev@kakao.com
