+++ src/pages/index.js
@@ -1,7 +1,7 @@
 import * as React from "react"
 import styled from "@emotion/styled"
 import Seo from "../components/seo"
-import Landing from "../components/common/Landing"
+import Landing from "../components/home/Landing"
 import Header from "../components/common/Header"
 import Footer from "../components/common/Footer"
 import GlobalStyle from "../components/common/GlobalStyle"
